Bug fix/check-user-verification

+++ src/services/commentService.js
@@ -25,7 +25,7 @@ async function getComments() {
     }
 }
 
-async function updateComment(commentId, body, countOfPluses, countOfMinuses) {
+async function updateComment(commentId, userId, body, countOfPluses, countOfMinuses) {
     try {
         //  Check if parameter are valid
         if(countOfPluses < 0 || countOfMinuses < 0) {
@@ -34,6 +34,11 @@ async function updateComment(commentId, body, countOfPluses, countOfMinuses) {
 
         //  Find comment of _id == commentId
         const comment = await Comment.findOne({ _id: commentId })
+
+        //  Check if user is owner of the comment
+        if(comment.userId != userId) {
+            throw new Error('User doesn\'t have permission to update comment!')
+        }
 
         //  Update comment
         comment.body = body
@@ -47,8 +52,16 @@ async function updateComment(commentId, body, countOfPluses, countOfMinuses) {
     }
 }
 
-async function deleteComment(commentId) {
+async function deleteComment(commentId, userId) {
     try {
+        //  Find comment of _id == commentId
+        const comment = await Comment.findOne({ _id: commentId })
+
+        //  Check if user is owner of the comment
+        if(comment.userId != userId) {
+            throw new Error('User doesn\'t have permission to update comment!')
+        }
+
         //  Delete comment of _id == commentId
         await Comment.deleteOne({ _id: commentId })
     } catch (err) {
